Miami Tasks for OSM TM
Based on Udarian's feedback, we will continue working on tasks in Miami but avoid areas that overlap with the community boundary. Udarian is anticipating to complete alignment in about 1-2 weeks, and then we will complete the tasks that are greyed out once alignment is completed. Changeset discussion here: https://www.openstreetmap.org/changeset/169617044